Add scaffolding for application
* I wrote a lot in Python, but I found it poorly ergonomic for
asynchronous and concurrent programming. While I could brute force my
way through writing code in it, I decided to try out Go. It's been a
while, but I'll give it a shot, and in the process, I think I'll learn a
ton.
* It's important to note that Go is not ideal for prototyping. Writing
code and getting immediate feedback is something you can get from Node
and Python, and maybe I'm regretting not choosing Node for this. (I
wrote onion routing on top of STREAM, but that's not really production
level code.)
* I wrote in Go to get as close as possible to production level code.
Hopefully I finish in time. Message passing concurrency, starting
goroutines, and understanding if I need a lock or not for some things is
where most of the confusion lies.

This commit:
* Includes API querying (not tested)
* Outline for the overall application, along with a basic socket example
that would handle multiple nodes.

Challenges:
* Getting a asynchronous server running alongside an asynchronous
client, all in one package! There's 0 good examples of this anywhere.
Everything separates client and server, and that's not ideal.
* Writing concurrent Go code, and getting over my urge to write safe
code for a system like this. Too much gut instinct toward making
production ready to overcome to ship faster.